move results scrapper code to results.py
